% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_clonotypes}
\alias{add_clonotypes}
\title{add clonotype information per cell}
\usage{
add_clonotypes(sobj, vdj_dirs, prefixes = NULL)
}
\arguments{
\item{sobj}{Seurat object}

\item{vdj_dirs}{path to clonotypes file}

\item{prefixes}{prefixes to add to cell_id}
}
\description{
add clonotype information per cell
}
