% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_features}
\alias{add_features}
\title{add features (i.e. rows) to a Seurat object
used for exporting various cell level measurements
for labeled in ucsc cell browser. Note that column names much be
identical to input seurat obj. Only the 'data' matrix will be altered}
\usage{
add_features(so, mat, prefix = NULL, assay = "RNA")
}
\arguments{
\item{so}{seurat object}

\item{mat}{data to add to expression matrix}

\item{prefix}{optional prefix to add to rownames of mat before}

\item{assay}{assay to which data will be appended, defaults to RNA
merging}
}
\description{
add features (i.e. rows) to a Seurat object
used for exporting various cell level measurements
for labeled in ucsc cell browser. Note that column names much be
identical to input seurat obj. Only the 'data' matrix will be altered
}
