% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_cellbrowser}
\alias{build_cellbrowser}
\title{Build cellbrowser}
\usage{
build_cellbrowser(
  dataset_paths,
  outdir = "cellbrowser_build",
  cbBuild_path = "/miniconda3/envs/py37/bin/cbBuild",
  command = TRUE,
  debug = FALSE
)
}
\arguments{
\item{dataset_paths}{vector of paths to cellbrowser directories}

\item{outdir}{output path}

\item{cbBuild_path}{path to cbBuild binary}

\item{command}{print command run}

\item{verbose}{if TRUE print debug messages}
}
\description{
Build cellbrowser
}
