% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da.R
\name{calc_da}
\alias{calc_da}
\title{Calculate differential abundance of cell count labels using edgeR}
\usage{
calc_da(
  sobj,
  cluster_col = NULL,
  sample_col = NULL,
  condition_col = NULL,
  custom_design = NULL
)
}
\arguments{
\item{sobj}{seurat object}

\item{cluster_col}{cluster or cell type column in meta.data}

\item{sample_col}{sample column in meta.data}

\item{condition_col}{column in meta.data indicating condition, defaults to sample_col if not
specified}

\item{custom_design}{design matrix for testing, defaults to `~condition_col`}
}
\description{
Calculate differential abundance of cell count labels using edgeR
}
\details{
See [Orchestraing Single Cell Analysis](https://osca.bioconductor.org/multi-sample-comparisons.html#differential-abundance)
}
