% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{closest_to_point}
\alias{closest_to_point}
\title{Get data  closest to a point}
\usage{
closest_to_point(mat, point)
}
\value{
Logical vector for each row in input matrix, TRUE if closest point
}
\description{
Get data  closest to a point
}
