% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsample_cells}
\alias{downsample_cells}
\title{Downsample seurat object to the same number of cells per ident}
\usage{
downsample_cells(so, group, n_cells = NULL, seed = 42)
}
\arguments{
\item{so}{seurat object}

\item{group}{grouping variable}

\item{n_cells}{# of desired cells, if null then downsample to minimum # of cells per group}

\item{seed}{integer for setting seed. if set to NULL then no seed will be set}
}
\description{
Downsample seurat object to the same number of cells per ident
}
