% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cell_count_matrix}
\alias{get_cell_count_matrix}
\title{Get cell counts from seurat object}
\usage{
get_cell_count_matrix(obj, row_var, col_var)
}
\arguments{
\item{row_var}{meta.data column to group for counts. will
be rows in the output matrix}

\item{col_var}{meta.data column to group for counts. will
be columns in the output matrix}

\item{sobj}{seurat object}
}
\description{
Get cell counts from seurat object
}
