% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_clonotypes}
\alias{get_clonotypes}
\title{get clonotype information per cell}
\usage{
get_clonotypes(vdj_dir, prefix = "")
}
\arguments{
\item{vdj_dir}{path to clonotypes file}

\item{prefix}{prefix to add to cell_id}
}
\description{
get clonotype information per cell
}
