% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_counts}
\alias{get_counts}
\title{Get cell counts from seurat object}
\usage{
get_counts(obj, ...)
}
\arguments{
\item{...}{grouping columns (quoted)}

\item{sobj}{seurat object}
}
\description{
Get cell counts from seurat object
}
