% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract out reduced dimensions and cell metadata to tibble}
\usage{
get_metadata(obj, embedding = NULL)
}
\arguments{
\item{obj}{Seurat Object}

\item{embedding}{dr slot to extract (defaults to all embeddings (2D))}
}
\description{
Extract out reduced dimensions and cell metadata to tibble
}
