% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jaccard_lists}
\alias{jaccard_lists}
\title{Compute similarities between lists of character vectors}
\usage{
jaccard_lists(x, y)
}
\arguments{
\item{x}{input list of character vectors}

\item{y}{input list of character vectors}
}
\value{
a matrix of pairwise jaccard coefficents
}
\description{
Compute similarities between lists of character vectors
}
