% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cellbrowser}
\alias{make_cellbrowser}
\title{Cell browser wrapper
Builds cell browser with better defaults}
\usage{
make_cellbrowser(
  so,
  column_list = NULL,
  primary_color_palette = discrete_palette_default,
  secondary_color_palette = palette_OkabeIto,
  secondary_cols = NULL,
  outdir = "cellbrowser",
  project = "seurat",
  marker_file = NULL,
  ident = "clusters",
  embeddings = names(so@reductions),
  color_skip = NULL,
  skip_expr_matrix = FALSE,
  skip_markers = FALSE,
  overwrite_cb_config = TRUE,
  annotate_markers = TRUE,
  cellbrowser_dir = "/miniconda3/bin/"
)
}
\arguments{
\item{so}{seurat object}

\item{column_list}{columns to keep in browser}

\item{primary_color_palette}{palette for catagorical variables}

\item{secondary_color_palette}{secondary palette for catagorical variables}

\item{secondary_cols}{columns to color by secondary_color_palette}

\item{outdir}{output directory for cellbrowser}

\item{project}{project string}

\item{marker_file}{seurat marker file path}

\item{ident}{default variable for labeling}

\item{embeddings}{embeddings to show in browser}

\item{color_skip}{no idea}

\item{skip_expr_matrix}{dont overwrite expression matrix}

\item{skip_markers}{dont overwrite markers}

\item{overwrite_cb_config}{overwritec cellbrowser.conf file}

\item{annotate_markers}{annotate markers using cellbrowser functionality}

\item{cellbrowser_dir}{directory where cellbrowser binary lives}
}
\description{
Cell browser wrapper
Builds cell browser with better defaults
}
