% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bc}
\alias{plot_bc}
\title{Plot barcode distribution}
\usage{
plot_bc(empty_drops, barcode_ranks, fdr = 0.01)
}
\arguments{
\item{empty_drops}{object produced by DropletUtils::emptyDrops}

\item{barcode_ranks}{object produced by DropletUtils::barcodeRanks}
}
\description{
Plot barcode distribution
}
