% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cell_proportions}
\alias{plot_cell_proportions}
\title{Plot cell proportions across each sample}
\usage{
plot_cell_proportions(
  obj,
  sample_id = "orig.ident",
  group_id = "coarse_clusters",
  facet_by = NULL,
  cols = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{sample_id}{column name containing the per sample id}

\item{group_id}{column name with cluster id}
}
\description{
Plot cell proportions across each sample
}
