% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_feature}
\alias{plot_feature}
\alias{plot_umap}
\alias{plot_tsne}
\alias{plot_pca}
\alias{plot_harmony}
\alias{plot_features_split}
\title{Plot cells in reduced dimensionality 2D space}
\usage{
plot_feature(
  seurat_obj,
  feature = NULL,
  plot_dat = NULL,
  pt_size = 0.001,
  pt_alpha = 1,
  label_text = FALSE,
  label_size = 6,
  label_color = NULL,
  .cols = NULL,
  cell_filter = NULL,
  palette_type = "cloupe",
  col_pal = "Reds",
  max_y = NULL,
  legend_title = NULL,
  embedding = "tsne",
  show_negative = FALSE,
  minimal_theme = FALSE,
  group = NULL,
  dims = c(1, 2),
  sorted = c("by_feature", "none", "random"),
  transform = "identity",
  na_col = "grey",
  ggrepel_opts = list(seed = 42, force = 1, max.overlaps = ncol(seurat_obj),
    segment.color = NA, size = label_size)
)

plot_umap(seurat_obj, ...)

plot_tsne(seurat_obj, ...)

plot_pca(seurat_obj, ...)

plot_harmony(seurat_obj, ...)

plot_features_split(
  seurat_obj,
  feature,
  group = "orig.ident",
  embedding = "umap",
  cols = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{feature}{feature to plot}

\item{plot_dat}{supplemental data.frame containing feature to plot.
Must have a column named cell that contains matching colnames in seurat_obj@data}

\item{pt_size}{size of points produced by geom_point}

\item{pt_alpha}{alpha value for points plotted by geom_point}

\item{label_text}{if TRUE display feature labels on plot}

\item{label_size}{size of label text}

\item{label_color}{color of label text, defaults to same colors as feature}

\item{.cols}{vector of colors to use for plot.}

\item{cell_filter}{character vector of cell names to include in plot}

\item{palette_type}{color palette type to use (either viridis, brewer, or cloupe)
defaults to using cellranger loupe-like colors}

\item{col_pal}{palette name to use if palette_type is brewer}

\item{max_y}{maximum feature value to set scale to. Defaults to max of the feature}

\item{legend_title}{string to supply for title for the legend}

\item{embedding}{dimensionality reduction to use for plotting}

\item{show_negative}{By default the legend value for continuous features will be clipped at zero.
If false, then the minumum value for the plotted feature will be used.}

\item{minimal_theme}{plot bare minimum}

\item{group}{grouping varible to split plots}

\item{dims}{which dims to plot from embedding, defaults to first and second, i.e. c(1,2).}

\item{sorted}{should the plotting be determined by sorting in ascending order? Default
is sorted by_feature (one of "by_feature", "none", "random")}

\item{transform}{adrgument o be passed to scale_color_gradientn for continuous data. defaults
to no transformation (i.e. "identity") See ?continous_scale for available transforms.}

\item{na_col}{Color for NA values (default = "grey")}

\item{ggrepel_opts}{named list of options to pass to ggrepel geom_text_repel.}

\item{\dots}{Additional parameters to pass to plot_feature}

\item{cols}{vector of cols to identity class, used to keep consistent colors
between plots}

\item{add_title}{want a title?}

\item{...}{additional args passed to plot_feature}
}
\description{
Cells can be colored by gene or feature in meta.data dataframe

See also plot_feature group argument
}
