% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_features_split}
\alias{plot_features_split}
\title{plot feature across multiple panels split by group}
\usage{
plot_features_split(
  sobj,
  feature,
  group = "orig.ident",
  embedding = "umap",
  cols = NULL,
  ...
)
}
\arguments{
\item{sobj}{seurat object}

\item{feature}{feature to plot}

\item{group}{grouping varible to split plots}

\item{embedding}{dimensionality reduction to use for plotting}

\item{cols}{vector of cols to identity class, used to keep consistent colors
between plots}

\item{...}{additional args passed to plot_feature}

\item{add_title}{want a title?}
}
\description{
plot feature across multiple panels split by group
}
