% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap per cluster/cell-type}
\usage{
plot_heatmap(
  obj,
  features,
  group,
  annotations = group,
  average = FALSE,
  slot = "scale.data",
  max_disp = 2.5,
  col_palettes = NULL,
  default_discrete_pal = discrete_palette_default,
  default_continuous_pal_fxn = viridis::inferno(256),
  hmap_options = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{features to plot as rows}

\item{group}{column from meta data to use to group cells across columns}

\item{annotations}{additional meta.data columns to add as column annotations
supplied as a character vector, defaults to display just group}

\item{average}{if TRUE, return average values per group rather than display all cells}

\item{slot}{data slot to retrive values defaults to scale.data}

\item{max_disp}{max/min value to display (2.5), only applied for scale.data}

\item{col_palettes}{list of alternative palettes to use for each annotation + group
variable supplied.}

\item{default_discrete_pal}{defatul discrete palette, defaults to scbp::discrete_palette_default,}

\item{hmap_options}{named list of options that are passed to ComplexHeatmap::Heatmap}

\item{normalize_cell_counts}{if TRUE, downsample cell counts to minimum cell count per group}

\item{default_continuous_pal}{default continuous color palette, defaults to viridis::inferno(256)}
}
\description{
Plot heatmap per cluster/cell-type
}
\examples{

plot_heatmap(Seurat::pbmc_small,
             features = rownames(Seurat::pbmc_small@assays$RNA@scale.data),
             group = "letter.idents",
             annotations = colnames(Seurat::pbmc_small@meta.data))

}
