% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_violins}
\alias{plot_violins}
\title{Make summary violin plots}
\usage{
plot_violins(
  seurat_obj,
  group,
  features,
  split_by = NULL,
  .size = 0.5,
  .width = 1,
  .scale = "width",
  .alpha = 1,
  cols = discrete_palette_default,
  rotate_x_text = TRUE,
  arrange_by_fill = TRUE,
  order_by_input = TRUE
)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{group}{x axis grouping variable}

\item{features}{features to plot}

\item{split_by}{additional metadata columns to split violins per group}

\item{.size}{violin size}

\item{.width}{violin width}

\item{.scale}{violin scaling parameter}

\item{.alpha}{violin alpha}

\item{cols}{color palette}

\item{rotate_x_text}{flip x and y}

\item{arrange_by_fill}{not sure}

\item{order_by_input}{maintain order of features supplied}
}
\description{
Make summary violin plots
}
