% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_alevin}
\alias{preprocess_alevin}
\title{Preprocessing for Alevin output}
\usage{
preprocess_alevin(paths, fdr = 0.01, cell_bc = NULL, genes = NULL, ...)
}
\arguments{
\item{paths}{paths to salmon output directory with alevin/quants_mat.gz files. If a
named vector is supplied then the cellbarcode will be
prefixed with "name_".}

\item{fdr}{fdr cutoff for cell calling}

\item{cell_bc}{list of character vectors of cell ids to keep in output,
if supplied cell calling will be skipped. Order must match paths argument}

\item{genes}{genes to use for cell calling, useful for exclude mito and ribogenes}

\item{...}{additional arguments passed onto emptyDrops}
}
\value{
A list of named lists with slots:
  mtx a sparseMatrix with only cell associated barcodes
  empty_drops a data.frame with information from emptyDrops
  barcode_ranks a data.frame with information from barcodeRanks
}
\description{
Preprocessing for Alevin output
}
