% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run fGSEA on gene lists}
\usage{
run_fgsea(
  ranked_gene_list,
  species = c("human", "mouse"),
  database = reactome.db::reactome.db,
  convert_ids = TRUE,
  min_size = 15,
  max_size = 500,
  n_perm = 10000,
  ...
)
}
\description{
Run fGSEA on gene lists
}
