% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbp-package.R
\docType{package}
\name{scbp-package}
\alias{scbp}
\alias{scbp-package}
\title{scbp: Boilerplate functions for single cell rna-seq analysis}
\description{
Lightweight wrappers to customize various plotting routines and
    analysis methods.
}
\author{
\strong{Maintainer}: Kent Riemondy \email{kent.riemondy@ucdenver.edu} (\href{https://orcid.org/0000-0003-0750-1273}{ORCID})

}
\keyword{internal}
