% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_shared_orthologs}
\alias{set_shared_orthologs}
\title{Compare two seurat objects from different species and return
matrices with one-to-one orthologs.}
\usage{
set_shared_orthologs(so1, so2, orthologs)
}
\arguments{
\item{so1}{suerat object}

\item{so2}{seurat object}

\item{orthologs}{data.frame with ortholog info}
}
\description{
Compare two seurat objects from different species and return
matrices with one-to-one orthologs.
}
