% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_avg_expr}
\alias{write_avg_expr}
\title{Write average expression matrix to a file}
\usage{
write_avg_expr(sobj, col, path, assay = "RNA")
}
\arguments{
\item{sobj}{seurat object}

\item{col}{metadata column for averaging}

\item{path}{output path}

\item{assay}{assay to write out, defaults to RNA}
}
\description{
Defaults to using write_tsv unless path ends with .csv or .csv.gz
}
