% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cellbrowser}
\alias{make_cellbrowser}
\title{Cell browser wrapper
Builds cell browser with better defaults}
\usage{
make_cellbrowser(
  so,
  column_list = NULL,
  primary_color_palette = discrete_palette_default,
  secondary_color_palette = palette_OkabeIto,
  secondary_cols = NULL,
  outdir = "cellbrowser",
  project = "seurat",
  marker_file = NULL,
  ident = "clusters",
  embeddings = names(so@reductions),
  color_skip = NULL,
  skip_expr_matrix = FALSE,
  skip_markers = FALSE,
  overwrite_cb_config = TRUE,
  annotate_markers = TRUE,
  default_assay = "RNA",
  assays = NULL,
  cellbrowser_dir = "/miniconda3/bin/",
  description = NULL,
  config = NULL,
  summary_html_format = "**{title}**  \\n{description}"
)
}
\arguments{
\item{so}{seurat object}

\item{column_list}{named list of columns to keep in browser. Names will be
displayed in the browser.}

\item{primary_color_palette}{palette for catagorical variables}

\item{secondary_color_palette}{secondary palette for catagorical variables}

\item{secondary_cols}{columns to color by secondary_color_palette}

\item{outdir}{output directory for cellbrowser}

\item{project}{project string}

\item{marker_file}{seurat marker file path}

\item{ident}{default variable for labeling}

\item{embeddings}{embeddings to show in browser}

\item{color_skip}{no idea}

\item{skip_expr_matrix}{dont overwrite expression matrix}

\item{skip_markers}{dont overwrite markers}

\item{overwrite_cb_config}{overwritec cellbrowser.conf file}

\item{annotate_markers}{annotate markers using cellbrowser functionality}

\item{default_assay}{assay to export (RNA)}

\item{assays}{character vector of additional assay expression matrices to export.
If supplied then the data matrix with be concatenated via rowbinding to the default assay matrix.
If a named character vector is supplied the name will be prefixed to the rows (i.e features).}

\item{cellbrowser_dir}{directory where cellbrowser binary lives}

\item{description}{named list with title and description fields to be rendered
into summary.html. Title will default to the project title.}

\item{config}{named list with custom entries to add to cellbrowser.conf file.
e.g. (config = list(priority = 1, radius = 5, alpha = 0.3)). See UCSC cellbrowser documentation for
allowable entries.}

\item{summary_html_format}{markdown formatted glue expression for formatting title and description into
summary.html file}
}
\description{
Cell browser wrapper
Builds cell browser with better defaults
}
