% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_bus}
\alias{preprocess_bus}
\title{Preprocessing for Bustools output}
\usage{
preprocess_bus(paths, fdr = 0.01, ...)
}
\arguments{
\item{paths}{paths to directories with mtx files. If a
named vector is supplied then the cellbarcode will be
prefixed with "name_".}

\item{fdr}{fdr cutoff for cell calling}

\item{...}{additional arguments passed onto emptyDrops}
}
\value{
A list of named lists with slots:
  mtx a sparseMatrix with only cell associated barcodes
  empty_drops a data.frame with information from emptyDrops
  barcode_ranks a data.frame with information from barcodeRanks
}
\description{
Preprocessing for Bustools output
}
