% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_markers_xlsx}
\alias{write_markers_xlsx}
\title{Write marker table to excel for easier perusal by collaborators}
\usage{
write_markers_xlsx(
  mrkr_list,
  path,
  description_string =
    "Genes differentially expressed between each cluster and all other cells",
  marker_type = NULL
)
}
\arguments{
\item{mrkr_list}{list of marker data.frames from Seurat or presto}

\item{path}{path to output xlsx}

\item{description_string}{String to add to first excel sheet to describe data
. Text is included with scbp::presto_marker_description or scbp::seurat_marker_description}

\item{marker_type}{one of c("Seurat", "presto", "scran_fm", "scran_sm"), if NULL,
will attempt to automatically identify}
}
\description{
Write marker table to excel for easier perusal by collaborators
}
